AAS_FallDamageDistance: long double->double drops 18 bytes (22->4)
Original IDA-decompiled types 'long double v0; float grav' caused MSVC6
to emit an extra fld+fstp pair (13 vs 12 insns). Using uniform 'double'
OUR+1 due to a remaining trailing fstp).

Source-commit: 637c368e0740ac51c8eea1581f8206c0bc59b7bb

+++ botlib/botlib.c
@@ -11908,8 +11908,8 @@ void __cdecl AAS_FaceCenter(int facenum, vec3_t center)
 //----- (10011520) --------------------------------------------------------
 __int64 AAS_FallDamageDistance()
 {
-  long double v0; // st7
-  float grav; // st6
+  double v0; // st7
+  double grav; // st6
 
   grav = libvar_sv_gravity->value;
   v0 = sqrt(300000.0) / grav;
